renamed table to table-cell in ui folder
