+++ front/src/modules/ui/table/table-cell/components/TableCellDisplayContainer.tsx
