+++ front/src/modules/ui/table/table-cell/components/TableCellDisplayMode.tsx
