+++ front/src/modules/ui/table/table-cell/components/TableCellEditMode.tsx
