+++ front/src/modules/ui/table/table-cell/components/TableCellSoftFocusMode.tsx
