+++ front/src/modules/ui/table/table-cell/hooks/useCurrentCellPosition.ts
