+++ front/src/modules/ui/table/table-cell/hooks/useCurrentTableCellEditMode.ts
