+++ front/src/modules/ui/table/table-cell/hooks/useIsSoftFocusOnCurrentTableCell.ts
