+++ front/src/modules/ui/table/table-cell/hooks/useSetSoftFocusOnCurrentTableCell.ts
