+++ front/src/modules/ui/table/table-cell/states/isCreateModeScopedState.ts
